Add constructor, rename to UnloadChunkRequestPacket, e
